refactor(gen-docs): split main.rs into focused modules
The 1500-line single-file generator had grown to bundle CLI
plumbing, syn-based extraction, type-graph analysis, serde-rename
logic, HTML rendering, markdown conversion, and syntect
highlighting. Split along the data-flow seam (extract → model →
render) with focused sub-modules inside each half:

- model.rs: Rule, ConfigDoc, ConfigField, TypeDoc, TypeKind,
  EnumVariant, StructField, Level, RenderContext.
- extract.rs: collect_rules, extract_rule, DeclareToolLint parser.
  - extract/config.rs: extract_config.
  - extract/ty.rs: toml_type_label, collect_referenced_idents,
    is_builtin_type, BUILTIN_TYPES, find_type_doc.
  - extract/serde_attrs.rs: serde_str_attr, apply_rename_all,
    pascal_to_snake, doc_attrs_to_markdown.
- render.rs: render_page, rule_article, level_badge, anchor_for,
  STYLE.
  - render/config.rs: config_section, custom_type_block.
  - render/markdown.rs: markdown_to_html, markdown_inline_to_html,
    highlight_code_blocks, highlight_to_html, lang_class_attr, and
    the SYNTAX_SET / THEME / HIGHLIGHT_CSS statics next to their
    only caller.

Tests move to co-located #[cfg(test)] mod tests blocks so each
helper's coverage lives next to the helper. No behaviour change:
all 13 existing tests pass and the rendered index.html is
byte-identical in a smoke run.
